Refactor data loading in streamlit_app.py
- Removed local CSV file paths and loading functions.
- Integrated data loading from the utils module for mechanical royalties and worldwide rates.
- Updated error handling messages for data loading failures.
